[android] jni binding refactor
Move registerNatives() call to a separate file to make jni.cpp
less dependent on native headers.
