+++ platform/android/core-files.json
@@ -38,6 +38,7 @@
         "platform/android/src/java/util.cpp",
         "platform/android/src/java_types.cpp",
         "platform/android/src/jni.cpp",
+        "platform/android/src/jni_native.cpp",
         "platform/android/src/logger.cpp",
         "platform/android/src/logging_android.cpp",
         "platform/android/src/map/camera_position.cpp",
@@ -142,6 +143,7 @@
         "java/util.hpp": "platform/android/src/java/util.hpp",
         "java_types.hpp": "platform/android/src/java_types.hpp",
         "jni.hpp": "platform/android/src/jni.hpp",
+        "jni_native.hpp": "platform/android/src/jni_native.hpp",
         "logger.hpp": "platform/android/src/logger.hpp",
         "map/camera_position.hpp": "platform/android/src/map/camera_position.hpp",
         "map/image.hpp": "platform/android/src/map/image.hpp",
