+++ platform/android/src/jni.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <string>
+#include <jni/jni.hpp>
 
 typedef struct _jmethodID* jmethodID;
 typedef struct _JavaVM JavaVM;
@@ -16,8 +17,6 @@ extern std::string dataPath;
 
 bool attach_jni_thread(JavaVM* vm, JNIEnv** env, std::string threadName);
 void detach_jni_thread(JavaVM* vm, JNIEnv** env, bool detach);
-
-extern void registerNatives(JavaVM* vm);
 
 } // namespace android
 } // namespace mbgl
