+++ platform/android/src/jni_native.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "jni.hpp"
+
+namespace mbgl {
+namespace android {
+
+void registerNatives(JavaVM* vm);
+
+} // namespace android
+} // namespace mbgl
