+++ platform/android/src/main.cpp
@@ -1,4 +1,5 @@
 #include "jni.hpp"
+#include "jni_native.hpp"
 #include <jni/jni.hpp>
 
 extern "C" JNIEXPORT jint JNI_OnLoad(JavaVM *vm, void *) {
@@ -6,4 +7,3 @@ extern "C" JNIEXPORT jint JNI_OnLoad(JavaVM *vm, void *) {
     mbgl::android::registerNatives(vm);
     return JNI_VERSION_1_6;
 }
-
